B23

+++ workbook/B23/atcoder4.cpp
@@ -1,6 +1,6 @@
 // g++ -std=c++11 -o atcoder atcoder.cpp
 // ./atcoder
-// 2026/8/29
+// 2026/8/28
 
 #include <algorithm>
 #include <cmath>
@@ -12,12 +12,14 @@ using namespace std;
 int main() {
   int N;
   cin >> N;
-  vector<int> X(N);
-  vector<int> Y(N);
+  vector<double> X(N);
+  vector<double> Y(N);
   for (int i = 0; i < N; i++) cin >> X[i] >> Y[i];
   int INF = 10e3;
   vector<vector<double>> DP((1 << N), vector<double>(N, INF));
+
   DP[1][0] = 0;
+
   for (int S = 0; S < (1 << N); S++) {
     for (int v = 0; v < N; v++) {
       if (DP[S][v] == INF) continue;
@@ -27,13 +29,14 @@ int main() {
 
         int dx = X[u] - X[v];
         int dy = Y[u] - Y[v];
-        double diff = sqrt(dx * dx + dy * dy);
+        int diff = sqrt(dx * dx + dy * dy);
 
         int nextS = S | (1 << u);
         DP[nextS][u] = min(DP[nextS][u], DP[S][v] + diff);
       }
     }
   }
+
   int all = (1 << N) - 1;
   double ans = INF;
   for (int v = 0; v < N; v++) {
